refactor: 패키지 구조 변경 및 컴포넌트 위치 이동
