+++ app/src/main/java/woowacourse/shopping/feature/cart/component/CartItem.kt
@@ -31,7 +31,7 @@ import androidx.compose.ui.unit.sp
 import woowacourse.shopping.R
 import woowacourse.shopping.feature.common.ProductQuantitySelector
 import woowacourse.shopping.feature.format.DecimalPriceFormatter
-import woowacourse.shopping.feature.productlist.PreviewableAsyncImage
+import woowacourse.shopping.feature.productlist.component.PreviewableAsyncImage
 
 @Composable
 fun CartItem(
