+++ app/src/main/java/woowacourse/shopping/feature/productlist/ProductListScreen.kt
@@ -22,8 +22,9 @@ import androidx.lifecycle.Lifecycle
 import androidx.lifecycle.compose.LifecycleEventEffect
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import androidx.lifecycle.viewmodel.compose.viewModel
-import woowacourse.shopping.feature.productlist.viewmodel.ProductListEvent
-import woowacourse.shopping.feature.productlist.viewmodel.ProductListViewModel
+import woowacourse.shopping.feature.productlist.component.ProductList
+import woowacourse.shopping.feature.productlist.component.ProductListAppBar
+import woowacourse.shopping.feature.productlist.component.RecentProductList
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
