+++ app/src/main/java/woowacourse/shopping/feature/productlist/component/LoadingIndicator.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.feature.productlist
+package woowacourse.shopping.feature.productlist.component
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.gestures.awaitEachGesture
