+++ app/src/main/java/woowacourse/shopping/feature/productlist/component/PreviewableAsyncImage.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.feature.productlist
+package woowacourse.shopping.feature.productlist.component
 
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.runtime.Composable
