+++ app/src/main/java/woowacourse/shopping/feature/productlist/component/RecentProductList.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.feature.productlist
+package woowacourse.shopping.feature.productlist.component
 
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Arrangement
