+++ app/src/test/java/woowacourse/shopping/feature/productlist/viewmodel/ProductListViewModelTest.kt
@@ -13,6 +13,7 @@ import woowacourse.shopping.feature.fake.FakeCartRepository
 import woowacourse.shopping.feature.fake.FakeProductRepository
 import woowacourse.shopping.feature.fake.FakeRecentProductRepository
 import woowacourse.shopping.feature.fake.FakeShoppingApplication
+import woowacourse.shopping.feature.productlist.ProductListViewModel
 
 @OptIn(ExperimentalCoroutinesApi::class)
 @ExtendWith(MainDispatcherExtension::class)
